setuptools build, avoid copying binaries when these are disabled

+++ cmake/cmake_extension.py
@@ -218,6 +218,9 @@ class BuildExtension(build_ext):
             if not src_file.is_file():
                 src_file = install_dir / ".." / (f + suffix)
 
+            if not src_file.is_file():
+                continue
+
             print(f"Copying {src_file} to {out_bin_dir}/")
             shutil.copy(f"{src_file}", f"{out_bin_dir}/")
